add multi-{key,query} stream to daml-ledger
This is an attempt to address #7034 & #7036. Strictly speaking, this
does not match their acceptance criteria, as this only supports multiple
queries, not mixed templates.

This also results in a somewhat confusing API because we end up with
overlapping functions: the new `streamQueries` can do everything that
the existing `streamQuery` can do, and the new `streamFetchByKeys` can
do everything that the existing `streamFetchByKey` can do, but we have
decided to keep the existing ones for now, to avoid breaking backwards
compatibility.

CHANGELOG_BEGIN

- [JavaScript Client Libraries] Two new methods have been added to
  `daml-ledger`: `streamQueries` and `streamFetchByKeys`. They are
  similar to the existing singular versions, except they can take
  multiple queries and multiple keys, respectively, and return a union
  of what the corresponding individual queries/keys would have.

CHANGELOG_END